--- src/databricks/sql/telemetry/enums/AuthMech.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class AuthMech(Enum):
-    OTHER = "other"
-    PAT = "pat"
-    OAUTH = "oauth"
