--- src/databricks/sql/telemetry/enums/DatabricksClientType.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class DatabricksClientType(Enum):
-    SEA = "SEA"
-    THRIFT = "THRIFT"
