--- src/databricks/sql/telemetry/enums/DriverVolumeOperationType.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class DriverVolumeOperationType(Enum):
-    TYPE_UNSPECIFIED = "type_unspecified"
-    PUT = "put"
-    GET = "get"
-    DELETE = "delete"
-    LIST = "list"
-    QUERY = "query"
