--- src/databricks/sql/telemetry/enums/StatementType.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class StatementType(Enum):
-    NONE = "none"
-    QUERY = "query"
-    SQL = "sql"
-    UPDATE = "update"
-    METADATA = "metadata"
