+++ src/databricks/sql/telemetry/telemetry_endpoint_models.py
@@ -4,6 +4,16 @@ from typing import List, Optional
 
 
 @dataclass
+class TelemetryRequest:
+    uploadTime: int
+    items: List[str]
+    protoLogs: Optional[List[str]]
+
+    def to_json(self):
+        return json.dumps(asdict(self))
+
+
+@dataclass
 class TelemetryResponse:
     errors: List[str]
     numSuccess: int
